new solution as per comments
